+++ vhost-user-backend/src/backend.rs
@@ -634,6 +634,10 @@ pub mod tests {
 
             backend
         }
+
+        pub fn events(&self) -> u64 {
+            self.events
+        }
     }
 
     impl VhostUserBackendMut for MockVhostBackend {