+++ vhost-user-backend/CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Deprecated
 ### Fixed
 
+- [[#343]](https://github.com/rust-vmm/vhost/pull/343) Avoid losing vring kicks
+
 ## v0.21.0
 
 ### Changed